add newton divide

+++ sem4/FundAlg/lab2/ex4/include/big_int.hpp
@@ -15,7 +15,7 @@ private:
     bool is_negative = false;
 
     void remove_leading_zeros();
-    static void remove_leading_zeros(std::vector<long long > & vector);
+    static void remove_leading_zeros(std::vector<long double> & vector);
     static bool is_correct_string(const std::string &str);
 
     static std::vector<std::complex<long double>> fft_transform(const std::vector<unsigned long long>& input, size_t n);
@@ -23,12 +23,12 @@ private:
 
     void shift_left(int k);
 
-    static std::vector<long long> big_int_to_polynomial(const BigInt &f);
+    static std::vector<long double> big_int_to_polynomial(const BigInt &f);
 
-    static BigInt polynomial_to_big_int(const std::vector<long long> &f, size_t base);
+    static BigInt polynomial_to_big_int(const std::vector<long double> &f, unsigned long long int base);
 
-    static std::vector<long long> polynomial_multiply(const std::vector<long long>& f, const std::vector<long long>& g);
-    static std::vector<long long> polynomial_subtract(const std::vector<long long>& f, const std::vector<long long>& g);
+    static std::vector<long double> polynomial_multiply(const std::vector<long double>& f, const std::vector<long double>& g);
+    static std::vector<long double> polynomial_subtract(const std::vector<long double>& f, const std::vector<long double>& g);
 
 
 public:
@@ -96,9 +96,9 @@ public:
     static BigInt karatsuba(BigInt f, BigInt g);
 
 
-    static std::vector<long long > newton_reverse(std::vector<long long > f, size_t l);
+    static std::vector<long double> newton_reverse(std::vector<long double> f, size_t l);
     static std::pair<BigInt, BigInt> newton_divide(const BigInt &lhs, const BigInt &rhs);
-    static std::vector<long long > reverse(std::vector<long long > f, size_t R);
+    static std::vector<long double> reverse(std::vector<long double> f, size_t R);
 
 
 };
